Remove friendlyErrorMessage circular reference
