+++ src/utils/protocol/__tests__/loadProtocol.test.js
@@ -6,7 +6,6 @@ import loadProtocol from '../loadProtocol';
 import protocolPath from '../protocolPath';
 import { readFile } from '../../filesystem';
 
-jest.mock('../../../ducks/modules/protocol');
 jest.mock('../../filesystem');
 jest.mock('../protocolPath');
